[JSONRPC] - Move traffic_ctl JSONRPC request/response definitions to a shared foler (#577)
* [JSONRPC] - Move JSONRPC request/response definitions to a shared folder.
- Split base request/response object into base objects and specific ones for traffic_ctl.
- Add documentation.
- Clean up some code in the RPC client code.

* [JSONRPC] - rpc shared client, add invoke function that takes a JSONRPC object.

* [JSONRPC] - fix typo

* [JSONRPC] - rename namespace. Add default template type.

* Fix typo in yaml_codecs.h
