+++ doc/developer-guide/jsonrpc/index.en.rst
@@ -29,5 +29,6 @@ JSONRPC
    jsonrpc-api.en
    jsonrpc-node.en
    jsonrpc-handler-development.en
+   jsonrpc-client-api.en
    traffic_ctl-development.en
    HandlerError.en
